made SocioIndustrialData its own class rather than subclass
